% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdj.R
\name{AddClonotype}
\alias{AddClonotype}
\title{Add 10X clonotype data to seurat object}
\usage{
AddClonotype(vdj.dir, scrna)
}
\arguments{
\item{vdj.dir}{String containing path to TCR (VDJ) directory.}

\item{scrna}{\linkS4class{Seurat} object.}
}
\value{
\linkS4class{Seurat} object with clonotype data (clonotype_id 
  and cdr3s_aa) added to the metadata for each cell.
}
\description{
\code{AddClonotype} adds clonotype info from matched 10X VDJ sequencing to 
the metadata of a given \linkS4class{Seurat} object.
}
\details{
This function is admittedly rough and will be rewritten in the future. It
does not include specific VDJ genes for each cell, rather just using the
final amino acid sequence for inter-sample comparison.
}
