% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{BatchCCEDA}
\alias{BatchCCEDA}
\title{Exploratory data analysis plots}
\usage{
BatchCCEDA(scrna, outdir, npcs = 50, batch = NULL)
}
\arguments{
\item{scrna}{Seurat object.}

\item{outdir}{Path to output directory for plots.}

\item{npcs}{Number of PCs to use for PCA and ElbowPlot. 50 by default.}

\item{batch}{String indicating \code{meta.data} column to be investigated for
batch effects.}
}
\value{
A Seurat object with a PCA for cell cycle genes stored with
  \code{reduction.name = "cc"}.
}
\description{
\code{BatchCCEDA} creates a number of plots to determine the number of PCs
to use for PCA/clustering and whether or not cell cycle scores and batch
effects should be addressed. Runs and plots an ElbowPlot to determine PCs
for later use. Runs and plots PCA for cell cycle genes to show their impact.
PCA on variable features can also be plotted by batch to view potential batch
effects.
}
