% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterdeg.R
\name{ClusterDEG}
\alias{ClusterDEG}
\title{Normalize, scale, and regress out wanted variation}
\usage{
ClusterDEG(scrna, outdir = ".", npcs = 30, res = 0.8, mnn = FALSE,
  skip.sct = FALSE, min.dist = 0.3, n.neighbors = 30,
  regress = NULL, ccpca = FALSE, test = "wilcox",
  logfc.thresh = 0.25, min.pct = 0.1)
}
\arguments{
\item{scrna}{\linkS4class{Seurat} object.}

\item{outdir}{Path to output directory.}

\item{npcs}{Number of principle components to use for UMAP and clustering.}

\item{res}{Numeric value denoting resolution to use for clustering. 
Higher values generally mean fewer clusters. Values of 0.5-3 are sensible. 
Multiple values may be entered as a vector - resulting clusters will be 
added as a \code{meta.data} column named \code{Cluster_res_npcs} where 
\code{res} and \code{npcs} will be the values for those arguments, 
respectively. The clusters derived from the last value in the list will be 
set as the default Ident for cells and stored in \code{meta.data} under
'seurat_clusters' in addition to the aforementioned format.}

\item{skip.sct}{Boolean indicating whether to skip 
\code{\link[Seurat]{SCTransform}}. Set to TRUE if 
\code{\link{SimpleIntegration}} was used to integrate the 
\code{Seurat} object.}

\item{min.dist}{Number that controls how tighly the embedding is allowed to
compress points together in \code{\link[Seurat]{RunUMAP}}. Increasing may 
be beneficial for large datasets.}

\item{n.neighbors}{Integer that determines the number of neighboring points
used in local approximations of manifold structure in 
\code{\link[Seurat]{RunUMAP}}. Values of 5-50 
are considered sensical. Larger values preserve more global structure while 
detailed local structure is lost.}

\item{regress}{Character vector of \code{meta.data} variables to regress 
during data scaling.}

\item{ccpca}{Boolean to indicate whether PCA using only cell cycle genes
should be done. If so, it will saved as a reduction named "cc". This is
useful to compare to prior PCAs using the cell cycle genes if cell cycle
scores were regressed out via \code{regress}.}

\item{test}{String indication which DE test to use for marker finding. 
Options are: 
  "wilcox", "bimod", "roc", "t", "negbinom", "poisson", "LR", 
  "MAST", "DESeq2". See \code{\link[Seurat]{FindAllMarkers}}.}

\item{logfc.thresh}{Value that limits DE testing to genes that show, 
on average, at least X-fold difference (log-scale) between two groups of 
cells. Increasing speeds up function at cost of potentially missing weaker 
differences.}

\item{min.pct}{Value that limits DE testing to genes detected in a minimum
fraction of cells in either population.}
}
\value{
A \linkS4class{Seurat} object with normalized, scaled counts and 
  assigned clusters. If \code{ccpca = TRUE}, an additional PCA reduction 
  named "cc" will also be present.
}
\description{
\code{ClusterDEG} runs \code{\link[Seurat]{SCTransform}} on a 
\linkS4class{Seurat} object, followed by \code{\link[Seurat]{RunPCA}}, 
\code{\link[Seurat]{RunTSNE}}, \code{\link[Seurat]{RunUMAP}}, and 
clustering. Also finds marker genes for each cluster and saves the output as
a table along with a heatmap of the top 10 upregulated genes in each cluster.
}
\details{
If multiple \code{res} values are given, a table and heatmap will be made for
each, along with saving the clusters for each in their own \code{meta.data} 
columns. 

Heatmaps created by \code{ClusterDEG} have each identity class downsampled 
to a max of 100 cells - this makes smaller clusters much more visible.
}
\examples{
\dontrun{
library(Seurat)
scrna <- ClusterDEG(pbmc_small)
}

\dontrun{
# Multiple clustering resolutions
scrna <- ClusterDEG(pbmc_small, res = c(0.8, 1, 1.2))
}

}
\author{
Jared Andrews
}
