% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{GetRootNodes}
\alias{GetRootNodes}
\title{Infer principal root node of each partition}
\usage{
GetRootNodes(cds, order.var, order.start, partitions = NULL)
}
\arguments{
\item{cds}{\code{cell_data_set} object containing a trajectory graph, as from
\code{\link{SeuratToCDS}}.}

\item{order.var}{String indicating \code{colData} column from which the 
time points or progression order is defined.}

\item{order.start}{String indicating variable in \code{order.col} that
represents the earliest time point or starting point.}

\item{partitions}{Character string indicating which partitions to calculate
root nodes for. Done for all partitions if not provided.}
}
\value{
Character vector containing the names of the root nodes for each
  partition. If \code{partitions} parameter is set, only the root nodes for
  those partitions will be returned.
}
\description{
\code{GetRootNodes} infers and returns the principal root node for each
partition. This will allow for the cells to be ordered according to the 
supplied variable (\code{order.var}) and start point \code{order.start}.
}
\author{
Jared Andrews
}
