% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{InferCellType}
\alias{InferCellType}
\title{Infer cell type using reference dataset}
\usage{
InferCellType(scrna, dataset, outdir, n_cores)
}
\arguments{
\item{scrna}{Seurat object.}

\item{dataset}{Path to tab-delimited table of gene counts. First column must
be gene identifiers that match those of the Seurat object. Each subsequent
column should contain counts for a cell type with the column header
denoting the cell type. Replicates should contain an underscore followed 
by the replicate number (e.g. BCell_1, BCell_2, etc).}

\item{outdir}{Path to output directory.}

\item{n_cores}{Number of cores to use for correlation. Linearly decreases 
computation time.}
}
\value{
A table containing the top three predicted cell types for each cell.
}
\description{
\code{InferType} utilizes a reference dataset to perform correlations of each
cell in a seurat object with each sample in the reference dataset. It retuns
a table containing the most likely cell type for each cell based on 
correlation values from the reference dataset.
}
\details{
The reference dataset can be from any source, it should just be normalized so
that columns (cell types) are comparable.
}
