% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{PerformCellInference}
\alias{PerformCellInference}
\title{Performs SingleR inference}
\usage{
PerformCellInference(scrna, sce, method, refset, labels, outdir,
  clusts.name = "NoClusters", clusts = NULL, singler.params = NULL)
}
\arguments{
\item{scrna}{\code{Seurat} object.}

\item{sce}{\code{SingleCellExperiment} object.}

\item{method}{String or character vector specifying whether annotation should 
be applied to each single cell \code{method = "single"} or aggregated into 
cluster-level profiles \code{method = "cluster"} prior to annotation. If
both are supplied, predictions will be performed for both methods.}

\item{refset}{Reference dataset as \linkS4class{SummarizedExperiment}.}

\item{labels}{String indicating whether to use broad lineage-based labels 
\code{labels="label.main"} for each reference sample or more fine-grained 
labels (\code{labels="label.fine"}.}

\item{outdir}{Path to output directory for annotation scores, distributions,
and heatmap.}

\item{clusts.name}{String indicating clusters column.}

\item{clusts}{Vector of cluster assignments for each cell.}

\item{singler.params}{List of arguments to be passed to 
\code{\link[SingleR]{SingleR}}.}
}
\value{
A \code{Seurat} object with inferred
  cell type information in the \code{meta.data} slot named in 
  \code{refset.labels} format. If \code{method = "cluster"}, the resulting
  \code{meta.data} column will be named in \code{clusters.refset.labels} 
  format.
}
\description{
\code{PerformCellInference} performs correlation-based cell type inference 
using a reference dataset via \code{\link[SingleR]{SingleR}}.
}
\author{
Jared Andrews
}
