% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{PrepRef}
\alias{PrepRef}
\title{Load and prepare reference dataset for cell type inference}
\usage{
PrepRef(scrna, dataset)
}
\arguments{
\item{scrna}{Seurat object.}

\item{dataset}{Path to tab-delimited table of gene counts. First column must
be gene identifiers that match those of the Seurat object. Each subsequent
column should contain counts for a cell type with the column header
denoting the cell type. Replicates contain an underscore followed by the
replicate number (e.g. BCell_1, BCell_2, etc).}
}
\value{
A list of two dataframes: "ref" - sorted genes from the reference 
  dataset also found in the seurat object, and "sc.sub" - sorted genes from 
  the seurat object also found in the reference dataset.
}
\description{
\code{PrepRef} processes a reference dataset of normalized gene counts for
correlation analysis with a seurat object. Removes unnecessary genes from the
reference dataset and matches the row ordering of the seurat object.
}
