% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{RunEnrichr}
\alias{RunEnrichr}
\title{Perform enrichment analyses for a gene list}
\usage{
RunEnrichr(genes, libraries = c("GO_Molecular_Function_2018",
  "GO_Cellular_Component_2018", "GO_Biological_Process_2018",
  "KEGG_2019_Human", "Reactome_2016", "BioCarta_2016", "Panther_2016"),
  outdir = NULL)
}
\arguments{
\item{genes}{A vector of gene symbols that will be tested.}

\item{libraries}{A vector of libraries to test the genes against.}

\item{outdir}{Path to output directory. NULL by default (tables will not be
saved, just returned).}
}
\value{
A named list of enrichment results for each library.
}
\description{
\code{RunEnrichr} submits gene lists to the enrichr web server for gene 
ontology/pathway enrichment analyses and collects the results as a named list
for each library used.
}
\details{
Potential libraries can be viewed with \code{listEnrichrDbs()} from the
\code{enrichR} package.

The enrichr web server can be found at <https://amp.pharm.mssm.edu/Enrichr/>.
If you use this function, you should cite the original authors as well as 
this package:
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/27141961}{Kuleshov MV, 
  Jones MR, Rouillard AD, et al. Enrichr: a comprehensive gene set enrichment 
  analysis web server 2016 update. Nucleic Acids Research. 2016; gkw377}
}
