% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{RunQC}
\alias{RunQC}
\title{Creates basic QC plots}
\usage{
RunQC(scrna, outdir = ".")
}
\arguments{
\item{scrna}{Seurat object.}

\item{outdir}{Path to output directory.}
}
\value{
Seurat object with percent mitochondrial reads added to the
  \code{meta.data} for each cell.
}
\description{
\code{RunQC} saves 3 QC plots showing gene counts, read counts, and percent 
mitochondrial reads per cell to help determine filters. It returns a
Seurat object with percent mitochondrial reads added to the \code{meta.data}.
}
\examples{
library(Seurat)
pbmc_small <- RunQC(pbmc_small)

}
