% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterdeg.R
\name{RunSCT}
\alias{RunSCT}
\title{Normalize, scale, and regress out wanted variation}
\usage{
RunSCT(scrna, outdir, npcs = 50, res = 0.8, min.dist = 0.3,
  n.neighbors = 30, regress = NULL, groups = NULL,
  groups.pca = NULL, groups.label = NULL, groups.legend = NULL,
  ccpca = FALSE, test = "wilcox", logfc.thresh = 0.25,
  min.pct = 0.1)
}
\arguments{
\item{scrna}{Seurat object.}

\item{outdir}{Path to output directory.}

\item{npcs}{Number of principle components to use for UMAP and clustering.
Default is 50, as \code{SCTransform} tends to do better with more.}

\item{res}{Numeric value denoting resolution to use for clustering. Default
is 0.8 (Seurat default). Increasing this value will speed up clustering,
but may decrease the numbers of distinct clusters. Values of 0.5-3 are 
sensible.}

\item{min.dist}{Number that controls how tighly the embedding is allowed to
compress points together in \code{RunUMAP}. Increasing may be beneficial
for large datasets. Default is 0.3 (Seurat default).}

\item{n.neighbors}{Integer that determines the number of neighboring points
used in local approximations of manifold structure in \code{RunUMAP}.
Altering it may be beneficial for large datasets (though it isn't stated
how it should be changed). Values of 5-50 are considered sensical.
Default is 30 (Seurat default).}

\item{regress}{Vector of metadata variables to regress during data scaling.
Must match column headers in metadata.}

\item{groups}{Vector of metadata variables to use for grouping in UMAP and 
PCA dimplots. Must match column headers in metadata.}

\item{groups.pca}{Vector of boolean values to determine if \code{DimPlots} 
PCA reductions should also be created for each group. NULL by default (
only UMAP reductions will be shown for each group). If provided, must be
same length as groups parameter.}

\item{groups.label}{Vector of boolean values to determine if \code{DimPlots} 
should show labels for each group. NULL by default (labels will not be 
shown). If provided, must be same length as groups parameter.}

\item{groups.legend}{Vector of boolean values to determine if \code{DimPlots} 
should show legends for each group. NULL by default (legends will be 
shown). If provided, must be same length as groups parameter.}

\item{ccpca}{Boolean to indicate whether PCA using only cell cycle genes
should be performed and plotted by sample identity and Phase.}

\item{test}{Denotes which DE test to use for marker finding. Options are: 
"wilcox" (default), "bimod", "roc", "t", "negbinom", "poisson", "LR", 
"MAST", "DESeq2".}

\item{logfc.thresh}{Value that limits DE testing to genes that show, 
on average, at least X-fold difference (log-scale) between two groups of 
cells. Increasing speeds up function at cost of potentially missing weaker 
signals. Default is 0.25 (Seurat default).}

\item{min.pct}{Value that limits DE testing to genes detected in a minimum
fraction of cells in either population. Default is 0.1 (Seurat default).}
}
\value{
A Seurat object with normalized, scaled counts and assigned clusters.
  If \code{ccpca = TRUE}, an additional PCA named "cc" will also be present.
}
\description{
\code{SCT} runs \code{SCTransform} on a Seurat object, followed by PCA, UMAP,
and clustering. Produces PCA and UMAP dimplots based on user-provided list
of metadata columns to use for grouping. Also finds marker genes and
saves the output as a table along with a heatmap of the top 10 upregulated
genes in each cluster.
}
