% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{ScoreAnnotatedMarkers}
\alias{ScoreAnnotatedMarkers}
\title{Score annotated marker lists}
\usage{
ScoreAnnotatedMarkers(scrna, marker.df)
}
\arguments{
\item{scrna}{\linkS4class{Seurat} object.}

\item{marker.df}{Dataframe with two columns named "Set" and "Marker". 
The "Set" column should contain a cell or process-type (e.g. Tcell, Bcell, 
Exhaustion markers, etc.) while the "Marker" column contains the 
comma-delimited gene symbols associated with it.}
}
\value{
\linkS4class{Seurat} object with scores for each "Set" added to 
  \code{meta.data}.
}
\description{
\code{ScoreAnnotatedMarkers} scores \linkS4class{Seurat} objects for lists of 
marker genes associated with a given annotation. Scores are stored in
\code{meta.data} columns named after each set.
}
\seealso{
\code{\link[Seurat]{AddModuleScore}} for scoring info and 
  \code{\link{VizAnnotatedMarkers}} for convenient visualization options.
}
