% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{SeuratToCDS}
\alias{SeuratToCDS}
\title{Convert a Seurat object to a cell data set object}
\usage{
SeuratToCDS(scrna, clusters, reduction = "pca")
}
\arguments{
\item{scrna}{A \linkS4class{Seurat} object with UMAP mappings.}

\item{clusters}{String indicating column in \code{meta.data} that 
contains cluster information to be utilized.}

\item{reduction}{String indicating which reduction should be used for 
feature loadings. Set to "mnn" if \code{method = "MNN"} was used with 
\link{SimpleIntegration}.}
}
\value{
A \code{cell_data_set} object with trajectory information calculated.
  Retains the same UMAP mappings and cluster information as \code{scrna}.
}
\description{
\code{SeuratToCDS} converts a \linkS4class{Seurat} object to a 
\code{cell_data_set} object with a trajectory graph.
}
\details{
This function preserves the UMAP coordinates and clustering from the
\code{Seurat} object. This means any plots created from the 
\code{cell_data_set} will be comparable to those from 
\code{\link[Seurat]{DimPlot}}.
}
\author{
Jared Andrews
}
