% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{SimpleIntegration}
\alias{SimpleIntegration}
\title{Integrate Seurat objects simply and easily}
\usage{
SimpleIntegration(scrnas, split.by = NULL, method = c("Seurat", "MNN"),
  vars.to.regress = NULL, n.features = 3000)
}
\arguments{
\item{scrnas}{Either a list of \linkS4class{Seurat} objects or a single 
\linkS4class{Seurat} object, the latter of which requires \code{split.by} 
to be provided as well.}

\item{split.by}{String containing name of meta.data column to be used to 
split Seurat object into multiple samples.}

\item{method}{String indicating method to be used. "Seurat" will use Seurat's
internal functions for SCT integration. "MNN" will use
\code{\link[SeuratWrappers]{RunFastMNN}}.}

\item{vars.to.regress}{Vector of meta.data variables to be regressed out
during \code{\link[Seurat]{SCTransform}}.}

\item{n.features}{Number of anchor features to use.}
}
\value{
An integrated \linkS4class{Seurat} object with technical 
  variation/batch effects removed from the individual samples.
}
\description{
\code{SimpleIntegration} integrates \linkS4class{Seurat} objects using 
either Seurat integration functions or 
\code{\link[SeuratWrappers]{RunFastMNN}}, a wrapper around
\code{\link[batchelor]{fastMNN}}. This is useful for removing batch effects. 
It can take either a list of \linkS4class{Seurat}  objects or a 
\code{meta.data} variable to split a single \linkS4class{Seurat} object by 
prior to integration.
}
\details{
Performs \code{\link[Seurat]{SCTransform}} on each Seurat object 
individually prior to integration.
}
\examples{
\dontrun{
library(Seurat)
# Will error due to too few cells.
pbmc.int <- SimpleIntegration(pbmc_small, split.by = "groups")
}

}
\seealso{

}
\author{
Jared Andrews
}
