% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{SplitIt}
\alias{SplitIt}
\title{Split lineages to get base lineages}
\usage{
SplitIt(x)
}
\arguments{
\item{x}{String to be split.}
}
\description{
For use with \code{lapply}. Returns only the first element of resulting list.
}
