% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizAnnotatedMarkers}
\alias{VizAnnotatedMarkers}
\title{Visualize an annotated marker list}
\usage{
VizAnnotatedMarkers(scrna, marker.file, outdir, pt.size = NULL,
  vln = NULL, ridge = NULL, dot = NULL, heatmap = NULL)
}
\arguments{
\item{scrna}{Seurat object.}

\item{marker.file}{Tab-delimited file with the first two columns named "Set"
and "Marker". The "Set" column should contain a cell or process-type (e.g.
Tcell, Bcell, Exhaustion markers, etc.) while the "Marker" column contains
the gene symbol associated with it. Many rows can be added to visualize 
many "Markers" for each "Set".}

\item{outdir}{Path to output directory.}

\item{pt.size}{Use to adjust point size for plotting FeaturePlots.}

\item{vln}{Boolean indicating whether to create Seurat VlnPlots for each set.
Splits by cell idents. NULL by default.}

\item{ridge}{Boolean indicating whether to create Seurat RidgePlots for each
set. Splits by cell idents. NULL by default.}

\item{dot}{Boolean indicating whether to create Seurat DotPlots for each
set. Splits by cell idents. NULL by default.}

\item{heatmap}{Boolean indicating whether to create a Seurat Heatmap for each
set. Splits by cell idents. NULL by default.}
}
\description{
\code{VizAnnotatedMarkers} creates Seurat FeaturePlots for lists of 
marker genes associated with a given annotation.
}
\details{
Plots for each annotation set will be output to a PDF that will be
dynamically sized and named based on the name and number of markers for the
set.
}
