% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizEnrichments}
\alias{VizEnrichments}
\title{Visualize GO/pathway enrichments}
\usage{
VizEnrichments(enrichments, outdir = "./", n.terms = 10,
  remove.insig = TRUE, adj.p.thresh = 0.05, colors = c("grey",
  "darkred"))
}
\arguments{
\item{enrichments}{Named list containing enrichment results as returned by
\code{\link{RunEnrichr}}}

\item{outdir}{Path to the output directory.}

\item{n.terms}{Number of terms to place on plot. 10 by default.}

\item{remove.insig}{Boolean indicating whether terms that don't meet the 
\code{adj.p.thresh} should be removed from the plots. TRUE by default.}

\item{adj.p.thresh}{Value indicating adjusted p-value threshold to filter 
terms. 0.05 by default.}

\item{colors}{Vector of two colors to use for coloring bars. First will be
low, second will be high. "grey" and "darkred" by default.}
}
\description{
\code{VizEnrichments} creates barcharts for enrichments returned by
\code{\link{RunEnrichr}}. Two plots will be created for each library - one 
ranked by adj. p-value, the other by Enrichr's combined score metrics. Plots 
will be saved in a PDF named by library.
}
\examples{
genes <- c("CD8A", "GZMA", "GZMK", "GZMB", "CD4", "CD3E", "GNLY")
libs <- c("Reactome_2016", "KEGG_2019_HUMAN")
terms <- RunEnrichr(genes, libraries = libs)

\dontrun{
VizEnrichments(enrichments = terms)
}

}
\seealso{
\code{\link{VizEnrichments}} for visualization.
}
