% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizRidgePlot}
\alias{VizRidgePlot}
\title{Creates RidgePlots for gene set}
\usage{
VizRidgePlot(scrna, outfile, genes)
}
\arguments{
\item{scrna}{Seurat object.}

\item{outfile}{Path for output PDF.}

\item{genes}{Vector of genes to plot.}
}
\description{
Dynamically generates Seurat RidgePlots for all genes in a given gene set as
a PDF.
}
