% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizScoredSets}
\alias{VizScoredSets}
\title{Visualize scored annotated marker sets}
\usage{
VizScoredSets(scrna, sets, outdir, pt.size = NULL, vln = NULL,
  ridge = NULL, dot = NULL, heatmap = NULL)
}
\arguments{
\item{scrna}{Seurat object.}

\item{sets}{Vector of}

\item{outdir}{Path to output directory.}

\item{pt.size}{Use to adjust point size for plotting FeaturePlots.}

\item{vln}{Boolean indicating whether to create Seurat VlnPlots for each set.
Splits by cell idents. NULL by default.}

\item{ridge}{Boolean indicating whether to create Seurat RidgePlots for each
set. Splits by cell idents. NULL by default.}

\item{dot}{Boolean indicating whether to create Seurat DotPlots for each
set. Splits by cell idents. NULL by default.}

\item{heatmap}{Boolean indicating whether to create a Seurat Heatmap for each
set. Splits by cell idents. NULL by default.}
}
\description{
\code{VizScoredSets} creates Seurat FeaturePlots for sets of 
marker genes associated with a given annotation.
}
\details{
Plots for each annotation set will be output to a PDF that will be
dynamically named based on the set name.
}
