% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizVDJDist}
\alias{VizVDJDist}
\title{Visualize clonotype distributions}
\usage{
VizVDJDist(scrna, outdir, g.by = NULL, o.by = NULL, n.clono.c = 10,
  n.clono.g = NULL)
}
\arguments{
\item{scrna}{Seurat object with clonotype data added to metadata with 
\code{AddClonotype}.}

\item{outdir}{Path to output directory.}

\item{g.by}{Metadata column to group samples by. If not provided, only
histograms of clonotypes will be saved.}

\item{o.by}{Vector containing names of members of each group to sort by 
within the group. Ignored if \code{g.by} is NULL. Should contain one 
instance of each potential value in \code{g.by} column if provided.}

\item{n.clono.c}{Number of top clonotypes to plot for comparison barchart.
Ignored if \code{g.by} is NULL.}

\item{n.clono.g}{Number of clonotypes to show in group-specific histograms.
All are shown by default.}
}
\description{
\code{VizVDJDist} visualizes clonotype distributions for each sample in a
seurat object as histograms as well as barcharts comparing clonotype 
proportions between them.
}
\details{
Cells with no clonotypes are still included in determining clonotype
frequencies, but NA is removed from subsequent graphs.
}
\author{
Jared Andrews
}
