% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{AssignCellType}
\alias{AssignCellType}
\title{Infers and assigns cell type for each cell using SingleR}
\usage{
AssignCellType(scrna, refset = c("hpca", "blueprint_encode", "immgen",
  "mouse.rnaseq"), labels = c("types", "main_types"),
  method = c("single", "cluster"), clusters = NULL, assign = TRUE,
  n.cores = 1, ...)
}
\arguments{
\item{scrna}{Seurat object.}

\item{refset}{String indicating which reference dataset to download and use
as the training set.}

\item{labels}{If \code{TRUE}, use broad lineage-based labels 
\code{main_types} for each reference sample. This is faster and can be more 
informative depending on how specific your data is. If \code{FALSE}, more
specific labels \code{types} will be used.}

\item{method}{String specifying whether annotation should be applied to 
each single cell \code{method = "single"} or aggregated into cluster-level 
profiles \code{method = "cluster"} prior to annotation.}

\item{clusters}{String defining the \code{meta.data} column in \code{scrna}
that specify cluster identities for each cell. Only used if \code{method 
= "cluster"}.}

\item{assign}{Boolean indicating whether inferred cell types should actually
be assigned to \code{Seurat} object or just returned as a \code{dataframe}.}

\item{n.cores}{Number of cores to use for correlation. Linearly decreases 
computation time.}

\item{...}{Additional arguments to be passed to 
\code{\link[SingleR]{SingleR}}.}
}
\value{
If \code{assign = TRUE}, returns a \code{Seurat} object with inferred
  cell type information in the \code{meta.data} slot named in 
  \code{refset.labels} format. If \code{method = "cluster"}, the resulting
  \code{meta.data} column will be named in \code{clusters.refset.labels} 
  format. If FALSE, returns a dataframe of the inferred cell type/lineage 
  information instead.
}
\description{
\code{AssignCellType} performs correlation-based cell inference using a 
reference dataset via \code{SingleR}.
}
\details{
Reference datasets available for use by this function via \code{SingleR} 
include:
\describe{
  \item{"hpca"}{Human Primary Cell Atlas (HPCA): A collection of Gene 
    Expression Omnibus (GEO datasets), which contains 713 microarray 
    samples classified to 38 main cell types and further annotated to 169
    subtypes.}
  \item{"blueprint_encode"}{Blueprint + ENCODE datasets: Blueprint 
    Epigenomics, 144 RNA-seq pure immune samples annotated to 28 cell 
    types. ENCODE, 115 RNA-seq pure stroma and immune samples annotated 
    to 17 cell types. Altogether, 259 samples with 43 cell types.}
  \item{"immgen"}{Immunological Genome Project (ImmGen): 830 microarray 
    samples, which we classified to 20 main cell types and further
    annotated to 253 subtypes.}
  \item{"mouse.rnaseq"}{A dataset of 358 mouse RNA-seq samples annotated to
    28 cell types. This dataset was collected, processed, and shared 
    courtesy of Bérénice Benayoun. This data set is especially useful for 
    brain-related samples.}
}
}
\examples{
preds <- AssignCellType(pbmc_small, refset = "hpca", labels = "main_types",
  method = "single")

preds2 <- AssignCellType(pbmc_small, refset = "hpca", labels = "types",
  method = "cluster", clusters = pbmc_small@meta.data$RNA_snn_res.1)

}
