% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{AssignCellType}
\alias{AssignCellType}
\title{Infers and assigns nearest cell type for each cell using SingleR}
\usage{
AssignCellType(scrna, refsets, labels = c("label.main", "label.fine"),
  outdir = NULL, method = c("single", "cluster"), clusters = NULL,
  singler.params = NULL)
}
\arguments{
\item{scrna}{\linkS4class{Seurat} object.}

\item{refsets}{List of reference dataset(s). Multiple can be provided.}

\item{labels}{String indicating whether to use broad lineage-based labels 
\code{labels="label.main"} for each reference sample or more fine-grained 
labels (\code{labels="label.fine"}. The former is quicker and can be 
informative enough if your sample has many cell types. The latter is 
best-suited for purified cell types or if particular cellular subtypes are 
important. If both are supplied, inference will be performed for both 
label sets.}

\item{outdir}{Path to output directory for annotation scores, distributions,
and heatmaps.}

\item{method}{String or character vector specifying whether annotation should 
be applied to each single cell \code{method = "single"} or aggregated into 
cluster-level profiles \code{method = "cluster"} prior to annotation. If
both are supplied, inference will be performed for both methods.}

\item{clusters}{String or character vector defining the \code{meta.data} 
column(s) in \code{scrna} that specify cluster identities for each cell. 
Only required if \code{method = "cluster"}. A character vector can be 
provided if multiple annotations with different clusters is wanted.

If provided with \code{method = "single"}, clusters will be used as an 
additional label in heatmap plotting.}

\item{singler.params}{List of additional arguments to be passed to 
\code{\link[SingleR]{SingleR}}.}
}
\value{
A \linkS4class{Seurat} object with inferred
  cell type information in the \code{meta.data} slot named in 
  \code{refset.labels} format. If \code{method = "cluster"}, the resulting
  \code{meta.data} column will be named in \code{clusters.refset.labels} 
  format. Pruned labels will also be added with \code{.pruned} appended to
  the column name.
}
\description{
\code{AssignCellType} performs correlation-based cell type inference using a 
reference dataset via \code{\link[SingleR]{SingleR}}.
}
\details{
Reference datasets available for use by this function include those provided
by SingleR:
\itemize{
  \item \code{\link[SingleR]{HumanPrimaryCellAtlasData}}
  \item \code{\link[SingleR]{BlueprintEncodeData}}
  \item \code{\link[SingleR]{DatabaseImmuneCellExpressionData}}
  \item \code{\link[SingleR]{NovershternHematopoieticData}}
  \item \code{\link[SingleR]{MonacoImmuneData}}
  \item \code{\link[SingleR]{ImmGenData}}
  \item \code{\link[SingleR]{MouseRNAseqData}}
}

If \code{outdir} is specified, the annotation results will be written to a
file named in \code{clusters.refset.labels.txt} format if 
\code{method="cluster"} or \code{refset.labels.txt} format if 
\code{method="single"}. Label distributions will be written to files named in
\code{refset.labels.dist.txt} and \code{refset.labels.pruned.dist.txt} 
format. Additionally, a heatmap will be made from the annotation results via 
\code{\link[SingleR]{plotScoreHeatmap}} if there are fewer than 65500 cells 
(the hclust method used fails with more cells than that).
}
\examples{
library("Seurat")
library("SingleR")
pbmc <- pbmc_small
# Download reference data from ExperimentHub.
hpca <- HumanPrimaryCellAtlasData()
# Always give the reference set a name.
metadata(hpca)$name <- "HPCA"
pbmc.pred <- AssignCellType(pbmc, refsets = list(hpca), 
  labels = "label.main", method = "single")

pbmc.clusters.pred <- AssignCellType(pbmc, refsets = list(hpca), 
  labels = "label.fine", method = "cluster", clusters = "RNA_snn_res.1")

\dontrun{
# Use both broad and fine labels, 'Single' and 'cluster' method, multiple 
# reference sets, and multiple cluster annotations.
bp <- BlueprintEncodeData()
metadata(bp)$name <- "Blueprint_Encode"
pbmc.all.anno <- AssignCellType(pbmc, refsets = list(hpca, bp), 
  clusters = c("RNA_snn_res.0.8", "RNA_snn_res.1"))
}

}
\seealso{
\code{\link[SingleR]{SingleR}} for additional options.
}
\author{
Jared Andrews
}
