% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{InferCellType}
\alias{InferCellType}
\title{Infer cell type using reference dataset}
\usage{
InferCellType(scrna, dataset, outdir, lineage = FALSE, n.cores = 1)
}
\arguments{
\item{scrna}{Seurat object.}

\item{dataset}{Path to tab-delimited table of gene counts. First column must
be gene identifiers that match those of the Seurat object. Each subsequent
column should contain counts for a cell type with the column header
denoting the cell type. Replicates should contain an underscore followed 
by the replicate number (e.g. Bcell_1, Bcell_2, etc).}

\item{outdir}{Path to output directory.}

\item{lineage}{Boolean indicating whether or not column names are formatted
as lineage followed by more granular specifications (e.g. Tcell.CD4_1,
Tcell.CD8_1, etc). If TRUE, will assign the lineage (everything before the)
first '.' in the column name to a metadata column called "base.lineage".
Columns not containing a '.' will use the entire name, though replicate
indicators will be removed (e.g. NK_1 will just be NK).}

\item{n.cores}{Number of cores to use for correlation. Linearly decreases 
computation time.}
}
\value{
A table containing the top three predicted cell types for each cell.
}
\description{
\code{InferType} utilizes a reference dataset to perform correlations of each
cell in a seurat object with each sample in the reference dataset. It retuns
a table containing the most likely cell type for each cell based on 
correlation values from the reference dataset.
}
\details{
The reference dataset can be from any source, it should just be normalized so
that columns (cell types) are comparable.
}
