% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{NormScoreCC}
\alias{NormScoreCC}
\title{Normalize counts and score cell cycle for each cell}
\usage{
NormScoreCC(scrna, skip.sct = NULL)
}
\arguments{
\item{scrna}{\code{Seurat} object to score cell cycle genes for each cell.}

\item{skip.sct}{Boolean indicating whether to skip \code{SCTransform} call.
Useful for integrated objects.}
}
\value{
Seurat object with cell cycle scores ('S.Score', 'G2M.Score') and 
  'Phase' added to \code{meta.data} for each cell.
}
\description{
\code{NormScoreCC} returns a \code{Seurat} object with normalized counts and 
adds cell cycle scores for each gene based on Seurat's cell cycle gene lists.
}
\examples{
\dontrun{
library(Seurat)
pbmc_small <- NormScoreCC(pbmc_small)
}

}
\author{
Jared Andrews
}
