% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{ScoreAnnotatedMarkers}
\alias{ScoreAnnotatedMarkers}
\title{Score annotated marker lists}
\usage{
ScoreAnnotatedMarkers(scrna, marker.df)
}
\arguments{
\item{scrna}{\linkS4class{Seurat} object.}

\item{marker.df}{Dataframe two columns named "Set" and "Marker". 
The "Set" column should contain a cell or process-type (e.g. Tcell, Bcell, 
Exhaustion markers, etc.) while the "Marker" column contains the 
comma-delimited gene symbols associated with it.}
}
\value{
\linkS4class{Seurat} object with scores for each "Set" added to 
  \code{meta.data}.
}
\description{
\code{ScoreAnnotatedMarkers} creates Seurat FeaturePlots for lists of 
marker genes associated with a given annotation.
}
\details{
Plots for each annotation set will be output to a PDF that will be
dynamically sized and named based on the name and number of markers for the
set. New directories will be created for each set in the output directory.
}
\seealso{
\code{\link[Seurat]{AddModuleScore}} for scoring info.
}
