% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{ScoreAnnotatedMarkers}
\alias{ScoreAnnotatedMarkers}
\title{Score annotated marker lists}
\usage{
ScoreAnnotatedMarkers(scrna, marker.df, assay = "SCT")
}
\arguments{
\item{scrna}{Seurat object.}

\item{marker.df}{Dataframe with the two columns called "Set" and "Marker". 
The "Set" column should contain a cell or process-type (e.g. Tcell, Bcell, 
Exhaustion markers, etc.) while the "Marker" column contains the 
comma-delimited gene symbols associated with it.}

\item{assay}{Assay to use tp snag expression for each gene. "SCT" is the 
preferred option, as "integrated" sets usually only have a few thousand
genes and "RNA" isn't normalized for any regressed variables.}
}
\value{
Seurat object with scores for each "Set" added to meta.data.
}
\description{
\code{ScoreAnnotatedMarkers} creates Seurat FeaturePlots for lists of 
marker genes associated with a given annotation.
}
\details{
Plots for each annotation set will be output to a PDF that will be
dynamically sized and named based on the name and number of markers for the
set. New directories will be created for each set in the output directory.
}
