% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{SimpleIntegration}
\alias{SimpleIntegration}
\title{Integrate Seurat objects simply and easily}
\usage{
SimpleIntegration(scrnas, split.by = NULL, skip.SCT = FALSE,
  vars.to.regress = NULL, n.features = 3000)
}
\arguments{
\item{scrnas}{Either a list of Seurat objects or a single Seurat object, the
latter of which requires \code{split.by} to be provided as well.}

\item{split.by}{String containing name of meta.data column to be used to 
split Seurat object into multiple samples.}

\item{skip.SCT}{Boolean indicating whether the \code{SCTransform} call for
each sample should be skipped or not. This should only be set to TRUE if
the samples if the samples have already undergone SCTransform.}

\item{vars.to.regress}{Vector of meta.data variables to be regressed out
during \code{SCTransform}.}

\item{n.features}{Number of anchor features to use. 3000 by default.}
}
\value{
An integrated Seurat object with technical variation/batch effects
  removed from the individual samples.
}
\description{
\code{SimpleIntegration} integrates Seurat objects using default Seurat 
parameters that should work for most cases. It can take either a list of 
Seurat objects or a meta.data variable to split a single Seurat object by
prior to integration.
}
\details{
Performs \code{SCTransform} on each Seurat object individually, per Seurat's
vignette for best practices
}
