% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizAnnotatedMarkers}
\alias{VizAnnotatedMarkers}
\title{Visualize an annotated marker list}
\usage{
VizAnnotatedMarkers(scrna, marker.df, outdir, vln = NULL, ridge = NULL,
  dot = NULL, heatmap = NULL, vln.params = NULL,
  ridge.params = NULL, dot.params = NULL, heatmap.params = NULL)
}
\arguments{
\item{scrna}{Seurat object.}

\item{marker.df}{Dataframe with the two columns called "Set" and "Marker". 
The "Set" column should contain a cell or process-type (e.g. Tcell, Bcell, 
Exhaustion markers, etc.) while the "Marker" column contains the 
comma-delimited gene symbols associated with it.}

\item{outdir}{Path to output directory.}

\item{vln}{Boolean indicating whether to create Seurat VlnPlots for each set.
Splits by cell idents. NULL by default.}

\item{ridge}{Boolean indicating whether to create Seurat RidgePlots for each
set. Splits by cell idents. NULL by default.}

\item{dot}{Boolean indicating whether to create Seurat DotPlots for each
set. Splits by cell idents. NULL by default.}

\item{heatmap}{Boolean indicating whether to create a Seurat Heatmap for each
set. Splits by cell idents. NULL by default.}

\item{vln.params}{List of keyword arguments to be passed to Seurat
\code{VlnPlot}. \code{features} are already defined and will throw
an error if passed.}

\item{ridge.params}{Boolean indicating whether to create Seurat RidgePlots 
for each set. Splits by cell idents. NULL by default.}

\item{dot.params}{Boolean indicating whether to create Seurat DotPlots for 
each set. Splits by cell idents. NULL by default.}

\item{heatmap.params}{Boolean indicating whether to create a Seurat Heatmap 
for each set. Splits by cell idents. NULL by default.}

\item{...}{Arguments to be passed to Seurat \code{FeaturePlot}. \code{cols}, 
\code{features}, \code{reduction}, and \code{ncol} are already defined and
will throw an error if passed.}
}
\description{
\code{VizAnnotatedMarkers} creates Seurat FeaturePlots for lists of 
marker genes associated with a given annotation.
}
\details{
Plots for each annotation set will be output to a PDF that will be
dynamically sized and named based on the name and number of markers for the
set. New directories will be created for each set in the output directory.
}
