% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizCellType}
\alias{VizCellType}
\title{Visualize inferred cell types}
\usage{
VizCellType(scrna, outdir, pt.size = NULL)
}
\arguments{
\item{scrna}{Seurat object with base.lineage, lineage, and celltype columns
added to the meta.data (as done by \code{AssignCellType}).}

\item{outdir}{Path to output directory.}

\item{pt.size}{Use to adjust point size for plotting.}
}
\description{
\code{VizCellType} creates Seurat DimPlots for the cell types and lineages
using both UMAP and TSNE reductions. These plots are automatically saved in
a PDF in the specified output directory.
}
