% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizDotPlot}
\alias{VizDotPlot}
\title{Creates DotPlot for gene set}
\usage{
VizDotPlot(scrna, outfile, genes, dot.params = NULL)
}
\arguments{
\item{scrna}{Seurat object.}

\item{outfile}{Path for output PDF.}

\item{genes}{Vector of genes to plot.}

\item{dot.params}{List of keyword arguments to be passed to 
\code{\link[Seurat]{DotPlot}}. \code{features} are already defined and will 
throw an error if passed.}
}
\description{
Dynamically generates a Seurat DotPlot for all genes in a given gene set as
a PDF.
}
