% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizHeatmap}
\alias{VizHeatmap}
\title{Creates Heatmap for gene set}
\usage{
VizHeatmap(scrna, outfile, genes, heatmap.params = NULL)
}
\arguments{
\item{scrna}{Seurat object.}

\item{outfile}{Path for output PDF.}

\item{genes}{Vector of genes to plot.}

\item{heatmap.params}{List of keyword arguments to be passed to 
\code{\link[Seurat]{DoHeatmap}}. \code{features} and \code{assay} are 
already defined and will throw an error if passed.}
}
\description{
Dynamically generates Seurat Heatmap for all genes in a given gene set as
a PDF.
}
\author{
Jared Andrews
}
