% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizMetaData}
\alias{VizMetaData}
\title{Visualize by metadata variables}
\usage{
VizMetaData(scrna, vars, outdir, ...)
}
\arguments{
\item{scrna}{Seurat object.}

\item{vars}{Vector of meta.data column names to plot.}

\item{outdir}{Path to output directory.}

\item{...}{Arguments to be passed to Seurat \code{DimPlot}. \code{cols}, 
\code{reduction}, and \code{group.by} are already defined and will throw
an error if passed.}
}
\description{
\code{VizDimMetaData} creates Seurat DimPlots for the given meta.data columns
using both UMAP, TSNE, and PCA reductions. These plots are automatically 
saved in a PDF in the specified output directory.
}
