% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{VizRidgePlot}
\alias{VizRidgePlot}
\title{Creates RidgePlots for gene set}
\usage{
VizRidgePlot(scrna, outfile, genes, ridge.params = NULL)
}
\arguments{
\item{scrna}{Seurat object.}

\item{outfile}{Path for output PDF.}

\item{genes}{Vector of genes to plot.}

\item{ridge.params}{List of keyword arguments to be passed to Seurat
\code{RidgePlot}. \code{features} are already defined and will throw
an error if passed. NULL by default.}
}
\description{
Dynamically generates Seurat RidgePlots for all genes in a given gene set as
a PDF.
}
